Create Lowest_Rating.py
Lowest RATING (100 Marks)
The Rating of N students is given with their name. You are asked to print the name of student(s) with lowest rating. You are required to use Nested Lists for the task.

Input Format
The first line of input consist of number of students, N.
There are 2 lines of input for each student. The first line consist of student name and second line consist of the rating.

Constraints
1<= Rating <=10

Output Format
Print the student(s) with lowest rating. If there is more than one student with lowest rating print the names of all in alphabetical order. 

Sample TestCase 1
Input
4
Rahul
1
Yogi
7
Modi
9
Gulam
1

Output
Gulam
Rahul