Create War_Between_Even_And_Odd.py
War between odd and even (100 Marks)
You will be given an array and you need to find odd numbers and add them, find even numbers and add them. Then compare the result, if the sum of even numbers found to be greater then print 'Even' to the stdout, if sum of odd number is greater than print 'odd' to the stdout. If sum of both found to be equal, then print 'Tied' to the stdout.

Input Format
You will be taking a number as an input stdin which tells about the length of the array. On another line, array elements should be there with single space between them.

Constraints
1 <= L <= 1000
1 <= Ai <= 1000

Output Format
You need to print 'Even', 'Odd' or 'Tied' according to the sum of odd and even numbers.

Sample TestCase 1
Input
9
74 32 31 91 77 88 96 44 23
Output
Even