Create Odd_Gets_Even_Bigger.py
For this challenge, you will be given an array and you are asked to form a number with the help of odd numbers available in the array. Actually, you need to form the bigger number out of the given odd numbers. In case you did not found any odd number, print 0 to the stdout.

Input Format
 You will be taking a number as an input stdin which tells about the length of the array. On another line, array elements should be there with single space between them.

Constraints
1 <= T <= 1000
1 <= Ai <= 1000

Output Format
Print the number that you will be forming with the odd numbers to the stdout.

Sample TestCase 1
Input
6
78 92 44 63 71 97
Output
977163
Explanation
In the given array, 63, 71 and 97 are the odd numbers and we need to form the biggest number using these numbers. So, the biggest number came out to be 977163.