Create BIggest_Difference.py
You will be given an array and you need to find the biggest difference between any two numbers present in the array.

Input Format
You will be taking a number as an input stdin which tells about the length of the array. On another line, array elements should be there with single space between them.

Constraints
1 <= L <= 1000
1 <= Ai <= 1000

Output Format
You need to print the biggest difference between any two numbers present in the array.

Note - The length of the array should not be less than 2.

Sample TestCase 1
Input
7
29 79 72 81 9 7 21
Output
74
Explanation
The biggest difference can only be found by find the smallest and the largest number in the array. Smallest number is 7 and the largest number is 81, so the biggest difference is 74.